+++ src/store/state.ts
@@ -43,6 +43,11 @@ export interface VQBState {
   selectedColumns: string[]; // focused in the DataViewer
 
   /**
+   * the right column names to be joined.
+   */
+  rightColumnNames: string[];
+
+  /**
    * object used to fill an edit step form
    */
   stepFormInitialValue?: object;
@@ -78,6 +83,7 @@ export function emptyState(): VQBState {
         totalCount: 0,
       },
     },
+    rightColumnNames: [],
     domains: [],
     currentStepFormName: undefined,
     stepFormInitialValue: undefined,