+++ tests/unit/join-step-form.spec.ts
@@ -15,7 +15,7 @@ describe('join Step Form', () => {
   runner.testCancel();
   runner.testResetSelectedIndex();
 
-  describe('AutocompleteWidget', () => {
+  describe('right dataset', () => {
     it('should instantiate an autocomplete widget with proper options from the store', () => {
       const initialState = {
         currentPipelineName: 'my_dataset',
@@ -43,7 +43,7 @@ describe('join Step Form', () => {
     });
   });
 
-  describe('ListWidget', () => {
+  describe('column names', () => {
     it('should pass down the "joinColumns" prop to the ListWidget value prop', async () => {
       const wrapper = runner.shallowMount(undefined, {
         data: {
@@ -87,6 +87,21 @@ describe('join Step Form', () => {
       });
       wrapper.find('listwidget-stub').vm.$emit('input', ['colRight', 'colLeft']);
       expect(wrapper.vm.$data.editedStep.on).toEqual(['colRight', 'colLeft']);
+    });
+
+    it('should fetch the right dataset columns when editing an existing step', () => {
+      runner.shallowMount(undefined, {
+        propsData: {
+          initialStepValue: {
+            name: 'join',
+            right_pipeline: 'pipeline_right',
+            type: 'left',
+            on: ['left_col', 'right_col'],
+          },
+        },
+      });
+      // Unfortunately, we can't easily mock the getColumnNamesFromPipeline action for now.
+      // So this test has no assertion.
     });
   });
 });